+++ managed/CounterStrikeSharp.Tests.Native/NativeTestsPlugin.cs
@@ -44,7 +44,8 @@ namespace NativeTestsPlugin
         {
             gameThreadId = Thread.CurrentThread.ManagedThreadId;
             // Loading blocks the game thread, so we use NextFrame to run our tests asynchronously.
-            Server.NextWorldUpdate(() => RunTests());
+            // Uncomment to run the tests on load
+            // Server.NextWorldUpdate(() => RunTests());
             AddCommand("css_run_tests", "Runs the xUnit tests for the native plugin.", (player, info) => { RunTests(); });
         }