+++ tooling/CodeGen.Natives/Scripts/GenerateGameEvents.cs
@@ -186,7 +186,8 @@ namespace CounterStrikeSharp.API.Core;
 
         foreach (var (gameEvent, definition) in events)
         {
-            File.WriteAllText(Path.Join(outputPath, $"Event{gameEvent.NamePascalCase}.g.cs"), template.Replace("<content>", definition));
+            File.WriteAllText(Path.Join(outputPath, $"Event{gameEvent.NamePascalCase}.g.cs"),
+                template.Replace("<content>", definition).ReplaceLineEndings("\r\n"));
         }
 
         Console.WriteLine($"Generated C# bindings for {allGameEvents.Count} game events successfully.");